pyproject.toml
[tool.poetry]
name = "kotatsu"
version = "0.1.0"
description = ""
authors = ["amaotone <amane.suzu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
slack-bolt = "^1.14.3"
fastapi = "^0.81.0"
uvicorn = "^0.18.3"
aiohttp = "^3.8.1"
python-dateutil = "^2.8.2"
pytest-diff = "^0.1.14"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
flake8 = "^5.0.4"
pytest = "^7.1.2"
pytest-mock = "^3.8.2"
freezegun = "^1.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest]
disable_test_id_escaping_and_forfeit_all_rights_to_community_support = true
